// src/app/enums/multiple-options-type.enum.ts
export enum MultipleOptionsType {
  DOCUMENTOS = 'DOCUMENTOS',
  RAP_ACCIONES = 'rap_acciones',
  RAP_APARATOS = 'rap_aparatos',
  RAP_AREAS_MEJORA = 'rap_areas_mejora',
  RAP_RESULTADOS = 'rap_resultados',
  RAP_CONTROL_HERMETICIDAD = 'rap_control_hermeticidad',
  RAP_CO_DILUIDO = 'rap_co_diluido',
  RAP_MEJORA_EQUIPOS = 'rap_mejora_equipos',
  RAP_SINGULARES = 'rap_singulares',
  RAP_UV = 'rap_uv',
}